FE-069: add a features page and link the footer to it
